feature: all board endpoints now require you to be an authorized user

+++ backend/app/routers/boards.py
@@ -2,9 +2,10 @@ from app.schemas import BoardBase, BoardCreate, BoardUpdate, BoardRead
 from fastapi import APIRouter, Depends, HTTPException, Response, status
 from app.crud import get_board_by_id, get_boards_by_owner_id, create_board, update_board, delete_board
 from app.database import get_db
+from app.models import User
 from sqlalchemy.orm import Session
 from collections.abc import Sequence
-
+from app.dependencies import get_current_user
 
 
 router = APIRouter(
@@ -18,18 +19,22 @@ router = APIRouter(
     "/{board_id}",
     response_model=BoardRead
 )
-def get_board_by_id_endpoint(board_id: int, db: Session = Depends(get_db)) -> BoardRead:
+def get_board_by_id_endpoint(board_id: int, db: Session = Depends(get_db), curent_user: User = Depends(get_current_user)) -> BoardRead:
     board = get_board_by_id(db, board_id)
     if not board:
         raise HTTPException(status_code=404, detail="Board not found")
+    if board.owner_id != curent_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to access this board")
     return BoardRead.model_validate(board)
 
 
 @router.get(
     "/owner/{owner_id}",
     response_model=Sequence[BoardRead])
-def get_boards_by_owner_id_endpoint(owner_id: int, db: Session = Depends(get_db)) -> Sequence[BoardRead]:
+def get_boards_by_owner_id_endpoint(owner_id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)) -> Sequence[BoardRead]:
     boards = get_boards_by_owner_id(db, owner_id)
+    if owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to access these boards")
     return [BoardRead.model_validate(board) for board in boards]
 
 
@@ -37,23 +42,32 @@ def get_boards_by_owner_id_endpoint(owner_id: int, db: Session = Depends(get_db)
     "/",
     response_model=BoardRead
     )
-def create_board_endpoint(board: BoardCreate, db: Session = Depends(get_db)) -> BoardRead:
+def create_board_endpoint(board: BoardCreate, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)) -> BoardRead:
     try:
-        db_board = create_board(db, board.title, board.description, board.owner_id)
+        db_board = create_board(db, 
+                                title=board.title, 
+                                description=board.description, 
+                                owner_id=current_user.id)
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
     return BoardRead.model_validate(db_board)
+
 
 @router.put(
     "/{board_id}",
     response_model=BoardRead
     )
-def update_board_endpoint(board_id: int, board: BoardUpdate, db: Session = Depends(get_db)) -> BoardRead:
+def update_board_endpoint(board_id: int, board: BoardUpdate, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)) -> BoardRead:
+    db_board = get_board_by_id(db, board_id)
+    if not db_board:
+        raise HTTPException(status_code=404, detail="Board not found")
+    if db_board.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to update this board")
     try:
-        db_board = update_board(db, board_id, board.title, board.description)
+        updated_board = update_board(db, board_id, title=board.title, description=board.description)
     except ValueError as e:
-        raise HTTPException(status_code=404, detail=str(e))
-    return BoardRead.model_validate(db_board)
+        raise HTTPException(status_code=400, detail=str(e))
+    return BoardRead.model_validate(updated_board)
 
 
 @router.delete(
@@ -61,7 +75,12 @@ def update_board_endpoint(board_id: int, board: BoardUpdate, db: Session = Depen
     status_code=status.HTTP_204_NO_CONTENT,
     response_model=None
     )
-def delete_board_endpoint(board_id: int, db: Session = Depends(get_db)) -> Response:
+def delete_board_endpoint(board_id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)) -> Response:
+    db_board = get_board_by_id(db, board_id)
+    if not db_board:
+        raise HTTPException(status_code=404, detail="Board not found")
+    if db_board.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to delete this board")
     try:
         delete_board(db, board_id)
     except ValueError as e:
